Split classes in separate files
